requirements.txt
numpy
pyaudio
soundfile
librosa
huggingface_hub[cli]
openai
faster-whisper
requests
ffmpeg
tiktoken
pyyaml
uvicorn
websockets
fastapi
pydantic_settings
starlette
sse_starlette
starlette_context
whisper_timestamped
elevenlabs
vllm
pynvml
bitsandbytes
flask
faiss-cpu
sentence-transformers
